notifier/requirements.txt
kafka-python
slack-sdk
python-snappy